# src/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "INVEBB API",
    "description": "INVEBB API Documentation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:10500",
      "description": "Development server"
    },
    {
      "url": "http://localhost:80",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/auth/info": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Sign up",
        "description": "",
        "parameters": [
          {
            "name": "userInfo",
            "in": "body",
            "description": "User info",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserInfo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully signed up"
          },
          "400": {
            "description": "Email is already registered"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "description": "",
        "parameters": [
          {
            "name": "userInfo",
            "in": "body",
            "description": "User info",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserInfo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userDetails": {
                      "$ref": "#/components/schemas/UserInfo"
                    }
                  },
                  "required": [
                    "userDetails"
                  ],
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userDetails": {
                      "$ref": "#/components/schemas/UserInfo"
                    }
                  },
                  "required": [
                    "userDetails"
                  ],
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password"
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshKey": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserInfo": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "abc@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "123456"
          }
        },
        "xml": {
          "name": "UserInfo"
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "example": "Error message"
          }
        },
        "xml": {
          "name": "ErrorResponse"
        }
      }
    }
  }
}